requirements.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0